added security to add, edit and delete views

+++ products/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.db.models.functions import Lower
 from django.contrib import messages
-
+from django.contrib.auth.decorators import login_required
 from .models import Product, Category
 from .forms import ProductForm
 
@@ -55,8 +55,13 @@ def product_description(request, product_id):
     return render(request, 'products/product_description.html', context)
 
 
+@login_required
 def add_product(request):
     '''Page management - add bake to page'''
+    if not request.user.is_superuser:
+        messages.error(request, 'For admin use only!')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -76,7 +81,12 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
+    if not request.user.is_superuser:
+        messages.error(request, 'For admin use only!')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     form = ProductForm(instance=product)
     messages.info(request, f'You are editing {product.name}')
@@ -101,7 +111,13 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
+
+@login_required
 def delete_product(request, product_id):
+    if not request.user.is_superuser:
+        messages.error(request, 'For admin use only!')
+        return redirect(reverse('home'))
+        
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
